+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { validate, mergeStyles, pluginGallerySchema } from 'wix-rich-content-common';
-import { isEqual } from 'lodash';
+import { isEqual, debounce } from 'lodash';
 import { convertItemData } from './lib/convert-item-data';
 import { DEFAULTS, isHorizontalLayout, sampleItems } from './constants';
 import resizeMediaUrl from './lib/resize-media-url';
@@ -10,6 +10,12 @@ import 'pro-gallery/dist/statics/main.min.css';
 import ExpandIcon from './icons/expand.svg';
 
 const { ProGallery } = process.env.SANTA ? {} : require('pro-gallery');
+
+function getClosestParanet(elem, selector) {
+  if (selector.some(selector => elem.matches(selector))) return elem;
+  if (elem === document) return null;
+  return getClosestParanet(elem.parentNode, selector);
+}
 
 class GalleryViewer extends React.Component {
   constructor(props) {
@@ -30,6 +36,18 @@ class GalleryViewer extends React.Component {
       });
     }
     window.addEventListener('resize', this.updateDimensions);
+    this.initUpdateDimensionsForDomChanges();
+  }
+
+  initUpdateDimensionsForDomChanges() {
+    const contentElement = getClosestParanet(this.container, ['.DraftEditor-root', '.viewer']);
+    this.observer = new MutationObserver(() => {
+      if (contentElement.clientHeight !== this.oldContentElementHeight) {
+        this.oldContentElementHeight = contentElement.clientHeight;
+        this.updateDimensions();
+      }
+    });
+    this.observer.observe(contentElement, { attributes: true, childList: true, subtree: true });
   }
 
   componentWillReceiveProps(nextProps) {
@@ -43,6 +61,7 @@ class GalleryViewer extends React.Component {
   }
 
   componentWillUnmount() {
+    this.observer.disconnect();
     window.removeEventListener('resize', this.updateDimensions);
   }
 
@@ -58,7 +77,7 @@ class GalleryViewer extends React.Component {
     }
   };
 
-  updateDimensions = () => {
+  updateDimensions = debounce(() => {
     if (this.container && this.container.getBoundingClientRect) {
       const width = Math.floor(this.container.getBoundingClientRect().width);
       let height;
@@ -67,7 +86,7 @@ class GalleryViewer extends React.Component {
       }
       this.setState({ size: { width, height } });
     }
-  };
+  }, 100);
 
   stateFromProps = props => {
     const items = props.componentData.items || DEFAULTS.items;